_codegen: add license and merge package 'imports'
Merge code from package internal/imports (recently copied from
package github.com/ernesto-jimenez/gogen/imports) into the codegen program itself.
Also copy the license attached to it into the _codegen/main.go source, as
_codegen/main.go was originally by the same author (Ernesto Jimenez) as
the imported code and matches the license terms of the global Testify
project.
Also: Ernesto, who was queried about this project, didn't block it.
https://github.com/stretchr/testify/pull/1782#discussion_r2322798422
